refactor: inline linkedin-client.ts into linkedin-create-post.ts
Only linkedinCreatePostNode consumed the client helpers, and 9 of 10
integrations in this repo inline HTTP/auth in their operation files
rather than extracting a *-client.ts module. Twitter is the sole
exception because OAuth 1.0a signing is non-trivial — LinkedIn's
bearer-token auth has no comparable complexity.

Moving all helpers into linkedin-create-post.ts as module-private
functions keeps the feature in one place and matches the project
norm. If a second LinkedIn operation arrives, the helpers can be
re-extracted at that point.

No public API or behavior changes. All 38 LinkedIn tests still pass.
